+++ pkg/azurecluster/validate_update.go
@@ -4,66 +4,29 @@ import (
 	"context"
 
 	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-	"k8s.io/api/admission/v1beta1"
 	capz "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
 	capiutil "sigs.k8s.io/cluster-api/util"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/azure-admission-controller/internal/errors"
 	"github.com/giantswarm/azure-admission-controller/internal/releaseversion"
 	"github.com/giantswarm/azure-admission-controller/internal/semverhelper"
 	"github.com/giantswarm/azure-admission-controller/pkg/generic"
-	"github.com/giantswarm/azure-admission-controller/pkg/validator"
+	"github.com/giantswarm/azure-admission-controller/pkg/key"
 )
 
-type UpdateValidator struct {
-	ctrlClient client.Client
-	logger     micrologger.Logger
-}
-
-type UpdateValidatorConfig struct {
-	CtrlClient client.Client
-	Logger     micrologger.Logger
-}
-
-func NewUpdateValidator(config UpdateValidatorConfig) (*UpdateValidator, error) {
-	if config.CtrlClient == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.CtrlClient must not be empty", config)
-	}
-	if config.Logger == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
-	}
-
-	v := &UpdateValidator{
-		ctrlClient: config.CtrlClient,
-		logger:     config.Logger,
-	}
-
-	return v, nil
-}
-
-func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.AdmissionRequest) error {
-	azureClusterNewCR := &capz.AzureCluster{}
-	azureClusterOldCR := &capz.AzureCluster{}
-	if _, _, err := validator.Deserializer.Decode(request.Object.Raw, nil, azureClusterNewCR); err != nil {
-		return microerror.Maskf(errors.ParsingFailedError, "unable to parse AzureCluster CR: %v", err)
-	}
-	if _, _, err := validator.Deserializer.Decode(request.OldObject.Raw, nil, azureClusterOldCR); err != nil {
-		return microerror.Maskf(errors.ParsingFailedError, "unable to parse AzureCluster CR: %v", err)
-	}
-
-	if !azureClusterNewCR.GetDeletionTimestamp().IsZero() {
-		a.logger.LogCtx(ctx, "level", "debug", "message", "The object is being deleted so we don't validate it")
-		return nil
-	}
-
-	capi, err := generic.IsCAPIRelease(azureClusterNewCR)
+func (h *WebhookHandler) OnUpdateValidate(ctx context.Context, oldObject interface{}, object interface{}) error {
+	azureClusterNewCR, err := key.ToAzureClusterPtr(object)
 	if err != nil {
 		return microerror.Mask(err)
 	}
-	if capi {
+	if !azureClusterNewCR.GetDeletionTimestamp().IsZero() {
+		h.logger.LogCtx(ctx, "level", "debug", "message", "The object is being deleted so we don't validate it")
 		return nil
+	}
+
+	azureClusterOldCR, err := key.ToAzureClusterPtr(oldObject)
+	if err != nil {
+		return microerror.Mask(err)
 	}
 
 	err = azureClusterNewCR.ValidateUpdate(azureClusterOldCR)
@@ -86,14 +49,10 @@ func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return microerror.Mask(err)
 	}
 
-	return a.validateRelease(ctx, azureClusterOldCR, azureClusterNewCR)
+	return h.validateRelease(ctx, azureClusterOldCR, azureClusterNewCR)
 }
 
-func (a *UpdateValidator) Log(keyVals ...interface{}) {
-	a.logger.Log(keyVals...)
-}
-
-func (a *UpdateValidator) validateRelease(ctx context.Context, azureClusterOldCR *capz.AzureCluster, azureClusterNewCR *capz.AzureCluster) error {
+func (h *WebhookHandler) validateRelease(ctx context.Context, azureClusterOldCR *capz.AzureCluster, azureClusterNewCR *capz.AzureCluster) error {
 	oldClusterVersion, err := semverhelper.GetSemverFromLabels(azureClusterOldCR.Labels)
 	if err != nil {
 		return microerror.Maskf(errors.ParsingFailedError, "unable to parse version from the AzureCluster being updated")
@@ -104,7 +63,7 @@ func (a *UpdateValidator) validateRelease(ctx context.Context, azureClusterOldCR
 	}
 
 	if !newClusterVersion.Equals(oldClusterVersion) {
-		cluster, err := capiutil.GetOwnerCluster(ctx, a.ctrlClient, azureClusterNewCR.ObjectMeta)
+		cluster, err := capiutil.GetOwnerCluster(ctx, h.ctrlClient, azureClusterNewCR.ObjectMeta)
 		if err != nil {
 			return microerror.Mask(err)
 		}
@@ -119,5 +78,5 @@ func (a *UpdateValidator) validateRelease(ctx context.Context, azureClusterOldCR
 		}
 	}
 
-	return releaseversion.Validate(ctx, a.ctrlClient, oldClusterVersion, newClusterVersion)
+	return releaseversion.Validate(ctx, h.ctrlClient, oldClusterVersion, newClusterVersion)
 }